test: merge astro-assets-prefix test files into one
